+++ config/config_testdata/benchmark/config_multipart_inject_500rps.json
@@ -1,0 +1,37 @@
+{
+    "steps": [
+      {
+        "id": 1,
+        "url": "https://testserver.ddosify.com/upload_image/",
+        "name": "",
+        "method": "POST",
+        "others": {
+          "h2": false,
+          "keep-alive": true,
+          "disable-redirect": true,
+          "disable-compression": false
+        },
+        "headers": {
+          "Content-Type": "multipart/form-data"
+        },
+        "timeout": 10,
+        "capture_env": {},
+        "payload_multipart": [
+          {
+            "src": "remote",
+            "name": "image",
+            "type": "file",
+            "value": "https://ddosify-backend-storage.s3.amazonaws.com/media/staging/multipart/vijzOnuqKiOkAWzAFo1wsGpdhk4M1DSGx99MIzPzufs/56_mv_1082778177_1681206249576109599_440.jpg"
+          },
+          {
+            "name": "ballot_id",
+            "value": "{{_randomInt}}"
+          }
+        ]
+      }
+    ],
+    "output": "stdout",
+    "duration": 10,
+    "load_type": "linear",
+    "iteration_count": 5000
+  }
